Get taxonomy from GBIF Backbone Taxonomy.

Retrieve accepted name taxonomy and retrieve taxonomic hierarchy for a df with a column of taxonomic names. Querying GBIF is done via the rgbif (Chamberlain and Oldoni 2021; Chamberlain and Boettiger 2017) function name_backbone.

get_gbif_tax(
  df,
  taxa_col = "original_name",
  out_file = tempfile(),
  king_type = "Plantae",
  do_common = FALSE,
  target_rank = "species"
)

Arguments

df
Dataframe with column of taxa names to resolve

taxa_col
Character. Name of column with taxa names

out_file
Character. Path to save results to. Saving is iterative as retrieving names can take some time.

king_type
Character. Kingdom type (i.e. Plantae, Animalia etc.)

do_common
Logical. If TRUE, also get common names. Takes much longer.

target_rank
Character. Default is 'species'. At what level of the taxonomic hierarchy are results desired.

Value

Dataframe of unique taxa names, other columns defining taxonomic hierarchy and, optionally, common names.

References

Chamberlain S, Boettiger C (2017). “R Python, and Ruby clients for GBIF species occurrence data.” PeerJ PrePrints. https://doi.org/10.7287/peerj.preprints.3304v1.

Chamberlain S, Oldoni D (2021). rgbif: Interface to the Global Biodiversity Information Facility API. R package version 3.6.0, https://CRAN.R-project.org/package=rgbif.

Examples

get_gbif_tax(data.frame(spp = "Eucalyptus viminalis"), taxa_col = "spp")
#> [1] "Eucalyptus viminalis"
#> Joining, by = "searched_name"
#> # A tibble: 1 x 26
#>   original_name  searched_name taxa  usageKey scientificName canonicalName rank 
#>   <chr>          <chr>         <chr>    <int> <chr>          <chr>         <chr>
#> 1 Eucalyptus vi~ Eucalyptus v~ Euca~  3177112 Eucalyptus vi~ Eucalyptus v~ SPEC~
#> # ... with 19 more variables: status <chr>, confidence <int>, matchType <chr>,
#> #   kingdom <chr>, phylum <chr>, order <chr>, family <chr>, genus <chr>,
#> #   species <chr>, kingdomKey <int>, phylumKey <int>, classKey <int>,
#> #   orderKey <int>, familyKey <int>, genusKey <int>, speciesKey <int>,
#> #   synonym <lgl>, class <chr>, stamp <dttm>