Knit child documents to report the standard envCleaned process

Use with knitr chunk option results="asis". Note: some files include named R objects to output the currently-used settings, and the knit will error if these are missing. For example, the 'core' clean files require the list settings with elements extent and grain, and settings_clean for the fbd_dist setting. Set these in the report environment before using report_clean.

knit_clean(clean_summary, heading_level = 3)

Arguments

clean_summary

Tibble containing names of clean process outputs (e.g. clean_start, bio_bin, bio_clean, etc), As output from envTargets::summarise_store_object()

heading_level

What level of subheading to set the output sections. e.g. default 3 will result in "###" heading level at the start of each clean Rmd

Value

as for knitr::knit_child()