Model the effect of principal components axes on taxa richness.

Model the effect of principal components axes on taxa richness.

make_effort_mod_pca(
  df,
  env_prcomp,
  context = "cell",
  threshold_lo = 0.05/2,
  threshold_hi = 0.05/2,
  effort_col = "qsize",
  effort_thresh = 3 * 3,
  ...
)

Arguments

df
Dataframe. Cleaned data specifying context.

env_prcomp
Output from env_pca.

context
Character. Column names that define context, usually a 'visit' to a 'cell'.

threshold_lo, threshold_hi
Numeric between 0 and 1 specifying the threshold above/below which richness is excessively above or below 'normal' and should be filtered.

effort_col
Character. Name of column with some measure of effort.

effort_thresh
Numeric. effort_col will be filtered below this threshold.

...
Passed to rstanarm::stan_glm().

Value

List of model outputs.