Find how taxa changed through the cleaning/filtering/tidying process

find_taxa does not work hierarchically. For example Eucalyptus will only match genus level records, not species records, such as Eucalyptus leucoxylon.

find_taxa(
  taxa,
  taxa_cols = c("original_name", "taxa"),
  lookup_taxa,
  filt_df_prefix = "flor_"
)

Arguments

taxa
Character. Taxa name to find.

taxa_cols
Character. Name of column(s) across data frames containing taxa information.

lookup_taxa
Dataframe with columns names matching taxa_cols.

filt_df_prefix
Character. Prefix used in each of the data frames created at each step in the filtering process.

Value

Examples

library(envClean)

# Example taxa
use_taxa <- "Eucalyptus gracilis"

# Set context
context <- c("lat", "long", "month", "year")

# Start
flor_start <- flor_all %>%
  tibble::as_tibble() %>%
  envFunc::add_time_stamp()

# Remove singletons
flor_single <- flor_start %>%
  filter_counts(context = context
                , thresh = 5
                ) %>%
  envFunc::add_time_stamp()

# Just keep most recent contexts
flor_recent <- flor_single %>%
  dplyr::group_by(across(any_of(context[!context %in% c("month", "year")]))) %>%
  dplyr::filter(year == max(year)
                , month == max(month)
                ) %>%
  dplyr::ungroup() %>%
  envFunc::add_time_stamp()

# Make sure required taxonomy objects are available
taxa <- make_taxa_taxonomy(df = data.frame(original_name = use_taxa))
#> [1] "1: Eucalyptus gracilis"
#> Joining, by = "searched_name"
#> Joining, by = "original_name"
#> Joining, by = c("taxa", "kingdom", "phylum", "order", "family", "genus",
#> "species", "class")

# Filter taxonomy
  # As lutaxa only contains `use_taxa`, this removes all other taxa
  # Normally, you'd have a row in lutaxa for every taxa
flor_taxa <- flor_recent %>%
  filter_taxa(context = context
              , taxonomy = taxa
              )
#> Joining, by = "original_name"
#> Joining, by = "taxa"
#> Joining, by = "original_name"

# How did records of 'taxa' change through the filtering?
find_taxa(taxa = use_taxa
          , lookup_taxa = taxa$lutaxa
          )
#> # A tibble: 0 x 6
#> # ... with 6 variables: name <chr>, obj <list>, has_stamp <lgl>, nrow <dbl>,
#> #   ctime <???>, summary <chr>