Get GBIF backbone taxonomy

Only queries GBIF for taxa not already in taxonomy_file.

get_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file = tempfile(),
  do_common = FALSE,
  remove_taxa = c("BOLD:", "dead", "unverified", "annual herb", "annual grass", "\\?"),
  remove_strings = c("\\sx\\s.*|\\sX\\s.*"),
  ...
)

Arguments

df

Dataframe with taxa column.

taxa_col

Character. Name of column with taxa names

taxonomy_file

Character. Path to save results to.

do_common

Logical. If true, an attempt will be made to find a common name.

remove_taxa

Character. Regular expressions to be matched. These will be filtered before searching.

remove_strings

Character. Regular expressions to be matched. These will be removed from the string before searching.

...

Arguments passed to rgbif::name_backbone_checklist().

Value

Dataframe. Results from envClean::get_gbif_tax(). Tweaked by column rank being lowercase and factor as per envClean::lurank.