Get taxonomy via galah::taxa_search()

Only queries galah for taxa not already in taxonomy_file. Can return a list, for several levels of taxonomic hierarchy, with the 'best' match at that level. For example, if 'genus' is provided in needed_ranks, the returned list will have an element 'genus' that contains, in a column named taxa, and for each of the original names provided, the best result at genus level or higher (in cases where no genus level match was available).

make_taxonomy(
  df = NULL,
  taxa_col = "original_name",
  taxonomy_file = tempfile(),
  force_new = list(original_name = NULL, timediff = as.difftime(26, units = "weeks")),
  remove_taxa = c("bold:", "BOLD:", "unverified", "annual herb", "annual grass",
    "incertae sedis", "\\?", "another species", "not naturalised in SA",
    "unidentified", "unverified", "annual tussock grass", "*no id"),
  remove_strings = c("\\sx\\s.*", "\\sX\\s.*", "\\s\\-\\-\\s.*",
    "\\ssp\\.$", "\\sssp\\.$", "\\sspec\\.$", "dead"),
  atlas = c("Australia"),
  return_taxonomy = TRUE,
  limit = TRUE,
  needed_ranks = c("species")
)

Arguments

df

Dataframe with taxa_col. Can be NULL only if taxonomy_file already exists.

taxa_col

Character or index. Name or index of column with taxa names. Each unique taxa in this column will appear in the results list element lutaxain a column called original_name

taxonomy_file

Character. File path to save results to. File type ignored. .parquet file used, partitioned on kingdom. If using this file directly (rather than via return_taxonomy = TRUE) remember to ungroup the result returned via arrow::open_dataset().

force_new

List with elements taxa_col and difftime. If taxonomy_file already exists any taxa_col matches between force_new and taxonomy_file will be requeried. Likewise any original_name that has not been searched since difftime will be requeried. Note the name taxa_col should be as provided as per the taxa_col argument. Set either to NULL to ignore.

remove_taxa

Character. Rows with regular expressions that match remove_taxa are removed (rows are removed).

remove_strings

Character. Text that matches remove_strings is removed from the string before searching (text, not row, is removed).

return_taxonomy

Logical. If TRUE, a list is returned containing the best match for each original_name in lutaxa and additional elements named for their rank (see envClean::lurank) with unique rows for that rank. One element per rank provided in needed_ranks

limit

Logical. If TRUE the returned list will be limited to those original_names in df

needed_ranks

Character vector of ranks required in the returned list. Can be "all" or any combination of ranks from envClean::lurank greater than or equal to subspecies.

Value

Null or list (depending on return_taxonomy). Writes taxonomy_file. If list, then elements:

raw - the 'raw' results returned from galah::search_taxa(), tweaked by column rank being an ordered factor as per envClean::lurank.

needed_ranks - One element for each rank specified in needed_ranks.

lutaxa - dataframe. For each unique name in taxa_col, the best taxa to use (taking into account each needed_ranks)

taxonomy - dataframe. For each taxa in lutaxa a row of taxonomic hierarchy

Details

Previous envClean::make_taxonomy() function is still available via envClean::make_gbif_taxonomy()

Examples

# setup
temp_file <- tempfile()
taxa_df <- tibble::tibble(names = c("blah", "Melithreptus gularis laetior", "Melithreptus gularis gularis", "Eucalyptus viminalis", "Eucalyptus viminalis cygnetensis"))

# make taxonomy (returns list and writes taxonomy_file)
taxonomy <- envClean::make_taxonomy(df = taxa_df, taxa_col = "names", taxonomy_file = temp_file, needed_ranks = c("kingdom", "genus", "species", "subspecies"))
#> Joining with `by = join_by(original_name)`
#> Warning: There were 2 warnings in `dplyr::filter()`.
#> The first warning was:
#> ℹ In argument: `is.na(as.numeric(gsub("[^[[:alnum:]]+", "", original_name)))`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> galah: version 2.0.2
#> ℹ Default node set to ALA (ala.org.au).
#> ℹ See all supported GBIF nodes with `show_all(atlases)`.
#> ℹ To change nodes, use e.g. `galah_config(atlas = "GBIF")`.
#> Matched 4 of 5 taxonomic search terms in selected atlas (Australia).
#> 1 unmatched search term:
#> • "blah"
#> 
#> saving results to H:/temp/nige\RtmpopTTUq\file629c3d9f198b.parquet
#> Joining with `by = join_by(original_name)`
#> Joining with `by = join_by(original_name)`
#> Joining with `by = join_by(original_name)`
#> Joining with `by = join_by(original_name)`
taxonomy$raw
#> # A tibble: 5 × 18
#>   original_name               search_term scientific_name scientific_name_auth…¹
#>   <chr>                       <chr>       <chr>           <chr>                 
#> 1 Eucalyptus viminalis        Eucalyptus… Eucalyptus vim… Labill.               
#> 2 Eucalyptus viminalis cygne… Eucalyptus… Eucalyptus vim… Boomsma               
#> 3 Melithreptus gularis gular… Melithrept… Melithreptus (… (Gould, 1837)         
#> 4 Melithreptus gularis laeti… Melithrept… Melithreptus (… Gould, 1875           
#> 5 blah                        blah        NA              NA                    
#> # ℹ abbreviated name: ¹​scientific_name_authorship
#> # ℹ 14 more variables: taxon_concept_id <chr>, rank <ord>, match_type <chr>,
#> #   kingdom <chr>, phylum <chr>, class <chr>, order <chr>, family <chr>,
#> #   genus <chr>, species <chr>, vernacular_name <chr>, issues <chr>,
#> #   stamp <dttm>, subspecies <chr>
taxonomy$kingdom
#> $lutaxa
#> # A tibble: 4 × 5
#>   original_name                    match_type matched_rank returned_rank taxa   
#>   <chr>                            <chr>      <ord>        <ord>         <chr>  
#> 1 Eucalyptus viminalis             exactMatch species      kingdom       Plantae
#> 2 Eucalyptus viminalis cygnetensis exactMatch subspecies   kingdom       Plantae
#> 3 Melithreptus gularis gularis     exactMatch subspecies   kingdom       Animal…
#> 4 Melithreptus gularis laetior     exactMatch subspecies   kingdom       Animal…
#> 
#> $taxonomy
#> # A tibble: 2 × 2
#>   taxa     kingdom 
#>   <chr>    <chr>   
#> 1 Plantae  Plantae 
#> 2 Animalia Animalia
#> 
taxonomy$genus
#> $lutaxa
#> # A tibble: 4 × 5
#>   original_name                    match_type matched_rank returned_rank taxa   
#>   <chr>                            <chr>      <ord>        <ord>         <chr>  
#> 1 Eucalyptus viminalis             exactMatch species      genus         Eucaly…
#> 2 Eucalyptus viminalis cygnetensis exactMatch subspecies   genus         Eucaly…
#> 3 Melithreptus gularis gularis     exactMatch subspecies   genus         Melith…
#> 4 Melithreptus gularis laetior     exactMatch subspecies   genus         Melith…
#> 
#> $taxonomy
#> # A tibble: 2 × 7
#>   taxa         kingdom  phylum     class         order         family      genus
#>   <chr>        <chr>    <chr>      <chr>         <chr>         <chr>       <chr>
#> 1 Eucalyptus   Plantae  Charophyta Equisetopsida Myrtales      Myrtaceae   Euca…
#> 2 Melithreptus Animalia Chordata   Aves          Passeriformes Meliphagid… Meli…
#> 
taxonomy$species
#> $lutaxa
#> # A tibble: 4 × 5
#>   original_name                    match_type matched_rank returned_rank taxa   
#>   <chr>                            <chr>      <ord>        <ord>         <chr>  
#> 1 Eucalyptus viminalis             exactMatch species      species       Eucaly…
#> 2 Eucalyptus viminalis cygnetensis exactMatch subspecies   species       Eucaly…
#> 3 Melithreptus gularis gularis     exactMatch subspecies   species       Melith…
#> 4 Melithreptus gularis laetior     exactMatch subspecies   species       Melith…
#> 
#> $taxonomy
#> # A tibble: 2 × 8
#>   taxa                 kingdom  phylum     class      order family genus species
#>   <chr>                <chr>    <chr>      <chr>      <chr> <chr>  <chr> <chr>  
#> 1 Eucalyptus viminalis Plantae  Charophyta Equisetop… Myrt… Myrta… Euca… Eucaly…
#> 2 Melithreptus gularis Animalia Chordata   Aves       Pass… Melip… Meli… Melith…
#> 
taxonomy$subspecies
#> $lutaxa
#> # A tibble: 4 × 5
#>   original_name                    match_type matched_rank returned_rank taxa   
#>   <chr>                            <chr>      <ord>        <ord>         <chr>  
#> 1 Eucalyptus viminalis             exactMatch species      species       Eucaly…
#> 2 Eucalyptus viminalis cygnetensis exactMatch subspecies   subspecies    Eucaly…
#> 3 Melithreptus gularis gularis     exactMatch subspecies   subspecies    Melith…
#> 4 Melithreptus gularis laetior     exactMatch subspecies   subspecies    Melith…
#> 
#> $taxonomy
#> # A tibble: 4 × 9
#>   taxa                kingdom phylum class order family genus species subspecies
#>   <chr>               <chr>   <chr>  <chr> <chr> <chr>  <chr> <chr>   <chr>     
#> 1 Eucalyptus viminal… Plantae Charo… Equi… Myrt… Myrta… Euca… Eucaly… NA        
#> 2 Eucalyptus viminal… Plantae Charo… Equi… Myrt… Myrta… Euca… Eucaly… Eucalyptu…
#> 3 Melithreptus (Eido… Animal… Chord… Aves  Pass… Melip… Meli… Melith… Melithrep…
#> 4 Melithreptus (Eido… Animal… Chord… Aves  Pass… Melip… Meli… Melith… Melithrep…
#> 

# query more taxa (results are added to taxonomy_file but only the new taxa are returned (default `limit = TRUE`)
more_taxa <- tibble::tibble(original_name = c("Amytornis whitei", "Amytornis striatus", "Amytornis modestus (North, 1902)", "Amytornis modestus modestus", "Amytornis modestus cowarie"))
taxonomy <- envClean::make_taxonomy(df = more_taxa, taxonomy_file = temp_file, needed_ranks = c("species"))
#> Joining with `by = join_by(original_name)`
#> Warning: There were 2 warnings in `dplyr::filter()`.
#> The first warning was:
#> ℹ In argument: `is.na(as.numeric(gsub("[^[[:alnum:]]+", "", original_name)))`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> saving results to H:/temp/nige\RtmpopTTUq\file629c3d9f198b.parquet
#> Joining with `by = join_by(original_name)`
taxonomy$species
#> $lutaxa
#> # A tibble: 5 × 5
#>   original_name                    match_type   matched_rank returned_rank taxa 
#>   <chr>                            <chr>        <ord>        <ord>         <chr>
#> 1 Amytornis modestus (North, 1902) canonicalMa… species      species       Amyt…
#> 2 Amytornis modestus cowarie       exactMatch   subspecies   species       Amyt…
#> 3 Amytornis modestus modestus      exactMatch   subspecies   species       Amyt…
#> 4 Amytornis striatus               exactMatch   species      species       Amyt…
#> 5 Amytornis whitei                 exactMatch   species      species       Amyt…
#> 
#> $taxonomy
#> # A tibble: 3 × 8
#>   taxa               kingdom  phylum   class order         family  genus species
#>   <chr>              <chr>    <chr>    <chr> <chr>         <chr>   <chr> <chr>  
#> 1 Amytornis modestus Animalia Chordata Aves  Passeriformes Maluri… Amyt… Amytor…
#> 2 Amytornis striatus Animalia Chordata Aves  Passeriformes Maluri… Amyt… Amytor…
#> 3 Amytornis whitei   Animalia Chordata Aves  Passeriformes Maluri… Amyt… Amytor…
#> 

# no dataframe supplied - all results in taxonomy_file returned
taxonomy <- envClean::make_taxonomy(taxonomy_file = temp_file, needed_ranks = c("subspecies"))
#> Joining with `by = join_by(original_name)`
taxonomy$subspecies
#> $lutaxa
#> # A tibble: 9 × 5
#>   original_name                    match_type   matched_rank returned_rank taxa 
#>   <chr>                            <chr>        <ord>        <ord>         <chr>
#> 1 Eucalyptus viminalis             exactMatch   species      species       Euca…
#> 2 Eucalyptus viminalis cygnetensis exactMatch   subspecies   subspecies    Euca…
#> 3 Melithreptus gularis gularis     exactMatch   subspecies   subspecies    Meli…
#> 4 Melithreptus gularis laetior     exactMatch   subspecies   subspecies    Meli…
#> 5 Amytornis modestus (North, 1902) canonicalMa… species      species       Amyt…
#> 6 Amytornis modestus cowarie       exactMatch   subspecies   subspecies    Amyt…
#> 7 Amytornis modestus modestus      exactMatch   subspecies   subspecies    Amyt…
#> 8 Amytornis striatus               exactMatch   species      species       Amyt…
#> 9 Amytornis whitei                 exactMatch   species      species       Amyt…
#> 
#> $taxonomy
#> # A tibble: 9 × 9
#>   taxa                kingdom phylum class order family genus species subspecies
#>   <chr>               <chr>   <chr>  <chr> <chr> <chr>  <chr> <chr>   <chr>     
#> 1 Eucalyptus viminal… Plantae Charo… Equi… Myrt… Myrta… Euca… Eucaly… NA        
#> 2 Eucalyptus viminal… Plantae Charo… Equi… Myrt… Myrta… Euca… Eucaly… Eucalyptu…
#> 3 Melithreptus (Eido… Animal… Chord… Aves  Pass… Melip… Meli… Melith… Melithrep…
#> 4 Melithreptus (Eido… Animal… Chord… Aves  Pass… Melip… Meli… Melith… Melithrep…
#> 5 Amytornis modestus  Animal… Chord… Aves  Pass… Malur… Amyt… Amytor… NA        
#> 6 Amytornis (Amytorn… Animal… Chord… Aves  Pass… Malur… Amyt… Amytor… Amytornis…
#> 7 Amytornis (Amytorn… Animal… Chord… Aves  Pass… Malur… Amyt… Amytor… Amytornis…
#> 8 Amytornis striatus  Animal… Chord… Aves  Pass… Malur… Amyt… Amytor… NA        
#> 9 Amytornis whitei    Animal… Chord… Aves  Pass… Malur… Amyt… Amytor… NA        
#> 

# clean up
rm(taxonomy)
unlist(temp_file)
#> [1] "H:/temp/nige\\RtmpopTTUq\\file629c3d9f198b"