Get taxonomy via galah::taxa_search()

Only queries galah for taxa not already in taxonomy_file. Can return a list, for several levels of taxonomic hierarchy, with the 'best' match at that level. For example, if 'genus' is provided in needed_ranks, the returned list will have an element 'genus' that contains, in a column named taxa, and for each of the original names provided, the best result at genus level or higher (in cases where no genus level match was available).

make_taxonomy(
  df,
  taxa_col = "original_name",
  taxonomy_file = tempfile(),
  force_new = list(original_name = NULL, timediff = as.difftime(26, units = "weeks")),
  remove_taxa = c("bold:", "BOLD:", "unverified", "annual herb", "annual grass",
    "incertae sedis", "\\?", "another species", "not naturalised in SA",
    "unidentified", "unverified", "annual tussock grass", "*no id"),
  remove_strings = c("\\sx\\s.*", "\\sX\\s.*", "\\s\\-\\-\\s.*",
    "\\ssp\\.$", "\\sssp\\.$", "\\sspec\\.$", "dead"),
  atlas = c("Australia"),
  return_taxonomy = TRUE,
  limit = TRUE,
  needed_ranks = c("species")
)

Arguments

df

Dataframe with taxa_col

taxa_col

Character or index. Name or index of column with taxa names. Each unique taxa in this column will appear in the results list element lutaxain a column called original_name

taxonomy_file

Character. File path to save results to. File type ignored. .parquet file used, partitioned on kingdom. If using this file directly (rather than via return_taxonomy = TRUE) remember to ungroup the result returned via arrow::open_dataset().

force_new

List with elements taxa_col and difftime. If taxonomy_file already exists any taxa_col matches between force_new and taxonomy_file will be requeried. Likewise any original_name that has not been searched since difftime will be requeried. Note the name taxa_col should be as provided as per the taxa_col argument. Set either to NULL to ignore.

remove_taxa

Character. Rows with regular expressions that match remove_taxa are removed (rows are removed).

remove_strings

Character. Text that matches remove_strings is removed from the string before searching (text, not row, is removed).

return_taxonomy

Logical. If TRUE, a list is returned containing the best match for each original_name in lutaxa and additional elements named for their rank (see envClean::lurank) with unique rows for that rank. One element per rank provided in needed_ranks

limit

Logical. If TRUE the returned list will be limited to those original_names in df

needed_ranks

Character vector of ranks required in the returned list. Can be "all" or any combination of ranks from envClean::lurank greater than or equal to subspecies.

Value

Null or list (depending on return_taxonomy). Writes taxonomy_file. If list, then elements:

needed_ranksOne element for each rank specified in needed_ranks

lutaxaDataframe. For each unique name in taxa_col, the best taxa to use (taking into account each needed_ranks)

taxonomyDataframe. For each taxa in lutaxa a row of taxonomic hierarchy

Details

Previous envClean::make_taxonomy() function is still available via envClean::make_gbif_taxonomy()

Examples

temp_file <- tempfile()
taxa_df <- tibble::tibble(name = c("blah", "Melithreptus gularis laetior", "Melithreptus gularis gularis", "Eucalyptus viminalis", "Eucalyptus viminalis cygnetensis"))
taxonomy <- make_taxonomy(df = taxa_df, taxonomy_file = temp_file, taxa_col = "name", needed_ranks = c("kingdom", "genus", "species", "subspecies"))
#> Error in df[taxa_col]: Can't subset columns that don't exist.
#> ✖ Column `name` doesn't exist.
taxonomy$raw
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
taxonomy$kingdom
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
taxonomy$genus
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
taxonomy$species
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
taxonomy$subspecies
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
more_taxa <- tibble::tibble(name = c("Amytornis whitei", "Amytornis striatus", "Amytornis modestus (North, 1902)", "Amytornis modestus modestus", "Amytornis modestus cowarie"))
taxonomy <- make_taxonomy(df = more_taxa, taxonomy_file = temp_file, taxa_col = "name", needed_ranks = c("species"))
#> Error in df[taxa_col]: Can't subset columns that don't exist.
#> ✖ Column `name` doesn't exist.
taxonomy$species
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
taxonomy <- make_taxonomy(df = more_taxa, limit = FALSE, taxonomy_file = temp_file, taxa_col = "name", needed_ranks = c("species"))
#> Error in df[taxa_col]: Can't subset columns that don't exist.
#> ✖ Column `name` doesn't exist.
taxonomy$species
#> Error in eval(expr, envir, enclos): object 'taxonomy' not found
rm(taxonomy)
#> Warning: object 'taxonomy' not found
unlist(temp_file)
#> [1] "H:/temp/nige\\RtmpMROTMz\\file6f887b4fa0"